Bank marketing campaigns data set analysis

It is a dataset that describing Portugal bank marketing campaigns results. Conducted campaigns were based mostly on direct phone calls, offering bank client to place a term deposit. If after all marking efforts client had agreed to place deposit - target variable marked 'yes', otherwise 'no'

Usage

data(bank)

Format

An object of class "data.frame"

job

Type of job

marital

marital status

education

education

default

has credit in default?

housing

has housing loan?

loan

has personal loan?

age

age

y

has the client subscribed a term deposit? (binary: "yes","no")

Details

Sourse of the data https://archive.ics.uci.edu/ml/datasets/bank+marketing

References

This dataset is public available for research. The details are described in S. Moro, P. Cortez and P. Rita. A Data-Driven Approach to Predict the Success of Bank Telemarketing. Decision Support Systems, Elsevier, 62:22-31, June 2014

Examples

data(bank)
head(bank_data)
#>         job marital   education default housing loan age  y
#> 1 housemaid married    basic.4y      no      no   no  56 no
#> 2  services married high.school unknown      no   no  57 no
#> 3  services married high.school      no     yes   no  37 no
#> 4    admin. married    basic.6y      no      no   no  40 no
#> 5  services married high.school      no      no  yes  56 no
#> 6  services married    basic.9y unknown      no   no  45 no