predict_outcome

Processing to predict response for various actual vs expected plots

Usage

predict_outcome(
  target,
  model_object,
  dataset,
  prediction_type = NULL,
  weights = NULL
)

Arguments

target

String of target variable name.

model_object

Model object. prettyglm currently supports

dataset

This is used to plot the number in each class as a barchart if plotly is TRUE.

prediction_type

type of prediction to be passed to the model object. For ...GLM defaults to ....

weights

weightings to be provided to predictions if required.

Value

dataframe

Returns a dataframe of Actual and Predicted Values

See also

Author

Jared Fowler